+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next
 * Diff-view for version history
 * Set form fields readonly instead of disabled
+* Update build system of Vue PDF rendering script from webpack to vite
 
 
 ## v2024.3 - 2024-01-09